Added extra info to README, added SSML knob, worked around a bug
The server now allows a client to post a ssml POST variable. This
disables any pre-conversion search/replace functions on the provided
text string.

Some (all?) SAPI voices, if the string begins with a "<", it is treated
as SSML. When ssml is true (non-zero), do nothing to the text, otherwise
replace < and > with "less than" and "greater than".

Also added a check of the length of the created audiostream, and if it
is 0, then return speech of the word "null" instead. This satisfies my
particular use-case. Other users may need to edit this to suit.

Added tts_sed() function to allow arbitrary search/replace or any other
operations on the passed text before it is converted to speech. Included
some commented out examples for emoticons.